work script

+++ .bash_functions/onepay.sh
@@ -17,9 +17,26 @@ get_repo_group() {
 
 get_repo_name() {
 	local repo_url=$(get_repo_url)
-	repo_url_https=$(echo "$repo_url" | sed -E 's#.*/([^/]+).git$#\1#')
-	echo $repo_url_https
+	repo_name=$(echo "$repo_url" | sed -E 's#.*/([^/]+).git$#\1#')
+	echo $repo_name
 }
+
+get_theme_name() {
+	local repo_group=$(get_repo_group)
+	if [[ "$repo_group" != "paygate" ]]; then
+		echo "CWD is not a theme"
+		return 1
+	fi
+
+	local repo_name=$(get_repo_name)
+	if [[ "$repo_name" == "paygate-general-fee" ]]; then
+		repo_name=$(basename "$PWD")
+	fi
+
+	local theme_name=$(echo "$repo_name" | sed -E 's#^paygate-(.*)$#\1#')
+	echo "$theme_name"
+}
+
 
 create_mr() {
     local source=""
@@ -95,17 +112,37 @@ replace_dev_domain() {
 }
 
 watch_angular() {
-	local theme="$1"
+	local repo_group=$(get_repo_group)
+	if [[ "$repo_group" != "paygate" ]]; then
+		echo "projects other than paygate/paygate-* are not implemented yet."
+		return 1
+	fi
+
+	theme=$(get_theme_name)
 	echo "Cleaning dist/..."
 	rm -rf dist/
-	echo "Starting watch build..."
-	ng build --configuration production --base-href=/paygate/${theme}/ --output-path=dist/paygate/${theme}/ --watch --watch
+
+	nvm use 16
+	local cmd="ng build --configuration production --base-href=/paygate/${theme}/ --output-path=dist/paygate/${theme}/ --watch"
+	echo "Running command:"
+	echo "$cmd"
+	eval "$cmd"
 }
 
 sync_theme() {
 	local host="$1"
 	local theme="$2"
-	rsync -av --progress --delete --rsh='ssh -p7602' dist/paygate/${theme}/ root@${host}:/usr/share/nginx/onepay.vn/paygate/${theme}/
+
+	if [[ -z "$host" || -z "$theme" ]]; then
+		echo "Usage: command <host> <theme>"
+		echo "Example: command dev18 installment"
+		return 1
+	fi
+
+	local cmd="rsync -av --progress --delete --rsh='ssh -p7602' dist/paygate/${theme}/ root@${host}:/usr/share/nginx/onepay.vn/paygate/${theme}/"
+	echo "Running command:"
+	echo "$cmd"
+	eval "$cmd"
 }
 
 build_theme_and_sync() {
@@ -115,8 +152,8 @@ build_theme_and_sync() {
 	set -e # Exit on error
 
 	if [[ -z "$host" || -z "$theme" ]]; then
-		echo "Usage: ${0} <host> <theme>"
-		echo "Example: ${0} dev18 installment"
+		echo "Usage: command <host> <theme>"
+		echo "Example: command dev18 installment"
 		return 1
 	fi
